chore: move hideAccount,showAccount to hook
